+++ BTCPayServer/Models/ManageViewModels/IndexViewModel.cs
@@ -10,6 +10,9 @@ namespace BTCPayServer.Models.ManageViewModels
         [MaxLength(50)]
         [Display(Name = "Email")]
         public string Email { get; set; }
+        [DataType(DataType.Password)]
+        [Display(Name = "Current password")]
+        public string CurrentPassword { get; set; }
         public bool EmailConfirmed { get; set; }
         public bool RequiresEmailConfirmation { get; set; }
         [Display(Name = "Name")]